+++ CHANGELOG.md
@@ -2724,3 +2724,4 @@
 - [2026-07-11 22:20] refactor(perf): optimize data flow between domain and data layers
 - [2026-07-11 22:29] feat(test): add bi-directional event bus handler for test
 - [2026-07-11 22:37] refactor(i18n): extract reusable atomic widgets from i18n
+- [2026-07-12 09:27] perf(analytics): apply RepaintBoundary on animated analytics widgets